feat: add singleton pattern class on base controller
